Add --database option
The new --database option allows one to import or sync into a secondary database, not the one that Drupal typically uses. This can be useful when you are preparing data to be migrated into Drupal, for example.

The Github editor kept messing up the indentation of the --concurrency option so I put it all on one line. Feel free to discard than hunk if desired.

The isDot() change is helpful to exclude stray files/dirs like .DS_Store and .git in the source directory.